settings: Move combine_code_language_options.
This commti moves combine_code_language_options to
settings_components.ts so that we can use it for stream
level setting as well.
